requirements.txt
# requirements.txt
discord.py>=2.3.2
uvicorn[standard]>=0.23.2
fastapi>=0.103.2
python-dotenv>=1.0.0
requests-oauthlib>=1.3.1
yt-dlp>=2023.7.6
PyNaCl>=1.5.0
aiohttp>=3.8.5
pydantic>=2.4.2
itsdangerous>=2.1.2
